[PATRO23-43] move methods to separate classes
