--- src/modules/budget/application/Data/DataHelperService.cs
@@ -1,286 +1,0 @@
-using System.Globalization;
-using Azure.Storage.Blobs;
-using Azure.Storage.Blobs.Models;
-using CsvHelper;
-using CsvHelper.Configuration;
-using Intive.Patronage2023.Modules.Budget.Application.Budget.CreatingBudget;
-using Intive.Patronage2023.Modules.Budget.Application.Budget.ExportingBudgets;
-using Intive.Patronage2023.Modules.Budget.Infrastructure.Data;
-using Intive.Patronage2023.Shared.Abstractions;
-using Intive.Patronage2023.Shared.Abstractions.Commands;
-using Intive.Patronage2023.Shared.Abstractions.Queries;
-using Intive.Patronage2023.Shared.Infrastructure.Domain.ValueObjects;
-using Intive.Patronage2023.Shared.Infrastructure.Domain;
-using Microsoft.AspNetCore.Http;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-
-namespace Intive.Patronage2023.Modules.Budget.Application.Data;
-
-/// <summary>
-/// Class DataService.
-/// </summary>
-public class DataHelperService
-{
-	private readonly IQueryBus queryBus;
-	private readonly ICommandBus commandBus;
-	private readonly IExecutionContextAccessor contextAccessor;
-	private readonly BlobServiceClient blobServiceClient;
-	private readonly BudgetDbContext budgetDbContext;
-
-	/// <summary>
-	/// Initializes a new instance of the <see cref="DataHelperService"/> class.
-	/// DataService.
-	/// </summary>
-	/// <param name="commandBus">The Command bus used for executing commands.</param>
-	/// <param name="queryBus">The Query bus used for executing queries.</param>
-	/// <param name="contextAccessor">The ExecutionContextAccessor used for accessing context information.</param>
-	/// <param name="configuration">The application's configuration, used for retrieving the connection string for the Blob Storage.</param>
-	/// <param name="budgetDbContext">The DbContext for accessing budget data in the database.</param>
-	public DataHelperService(IQueryBus queryBus, ICommandBus commandBus, IExecutionContextAccessor contextAccessor, IConfiguration configuration, BudgetDbContext budgetDbContext)
-	{
-		this.queryBus = queryBus;
-		this.commandBus = commandBus;
-		this.contextAccessor = contextAccessor;
-		this.blobServiceClient = new BlobServiceClient(configuration.GetConnectionString("BlobStorage"));
-		this.budgetDbContext = budgetDbContext;
-	}
-
-	/// <summary>
-	/// Checks if a blob container exists, and if not, creates one.
-	/// </summary>
-	/// <param name="containerName">The name of the container to be checked/created.</param>
-	/// <returns>A client reference to the newly created or existing blob container.</returns>
-	public async Task<BlobContainerClient> CreateBlobContainerIfNotExists(string containerName)
-	{
-		var containerClient = this.blobServiceClient.GetBlobContainerClient(containerName);
-		await containerClient.CreateIfNotExistsAsync();
-		return containerClient;
-	}
-
-	/// <summary>
-	/// Validates the properties of a budget object.
-	/// </summary>
-	/// <param name="budget">The budget object to validate.</param>
-	/// <returns>A list of error messages. If the list is empty, the budget object is valid.</returns>
-	public List<string> ValidateBudget(GetBudgetsToExportInfo budget)
-	{
-		var errors = new List<string>();
-
-		if (string.IsNullOrEmpty(budget.Name))
-		{
-			errors.Add("Budget name is missing");
-		}
-
-		if (string.IsNullOrEmpty(budget.IconName))
-		{
-			errors.Add("Budget icon name is missing");
-		}
-
-		if (string.IsNullOrEmpty(budget.Currency))
-		{
-			errors.Add("Budget currency is missing");
-		}
-
-		if (string.IsNullOrEmpty(budget.Value))
-		{
-			errors.Add("Budget value is missing");
-		}
-		else if (!decimal.TryParse(budget.Value, NumberStyles.Any, CultureInfo.InvariantCulture, out _))
-		{
-			errors.Add("Budget value is not a valid decimal number");
-		}
-
-		if (string.IsNullOrEmpty(budget.StartDate))
-		{
-			errors.Add("Budget start date is missing");
-		}
-		else if (!DateTime.TryParse(budget.StartDate, CultureInfo.InvariantCulture, DateTimeStyles.None, out _))
-		{
-			errors.Add("Budget start date is not a valid date");
-		}
-
-		if (string.IsNullOrEmpty(budget.EndDate))
-		{
-			errors.Add("Budget end date is missing");
-		}
-		else if (!DateTime.TryParse(budget.EndDate, CultureInfo.InvariantCulture, DateTimeStyles.None, out _))
-		{
-			errors.Add("Budget end date is not a valid date");
-		}
-
-		if (!string.IsNullOrEmpty(budget.StartDate) && !string.IsNullOrEmpty(budget.EndDate) &&
-			DateTime.Parse(budget.StartDate) >= DateTime.Parse(budget.EndDate))
-		{
-			errors.Add("Budget start date cannot be later than or equal to end date");
-		}
-
-		return errors;
-	}
-
-	/// <summary>
-	/// Retrieves all budgets to be exported.
-	/// </summary>
-	/// <returns>A list of budget objects to be exported. Returns null if no budgets are found.</returns>
-	public async Task<List<GetBudgetsToExportInfo>?> GetBudgetsToExport()
-	{
-		var query = new GetBudgetsToExport() { };
-		return await this.queryBus.Query<GetBudgetsToExport, List<GetBudgetsToExportInfo>?>(query);
-	}
-
-	/// <summary>
-	/// Generates a local file path for a new CSV file.
-	/// </summary>
-	/// <returns>A string representing the local path to a newly generated CSV file.</returns>
-	public string GenerateLocalCsvFilePath()
-	{
-		string localPath = "data"; ////src\api\app\data
-		Directory.CreateDirectory(localPath);
-		string fileName = Guid.NewGuid().ToString() + ".csv";
-		return Path.Combine(localPath, fileName);
-	}
-
-	/// <summary>
-	/// Writes a list of budgets to a CSV file at the specified file path.
-	/// </summary>
-	/// <param name="budgets">A list of budgets to be written to the CSV file.</param>
-	/// <param name="filePath">The local path of the CSV file.</param>
-	/// <returns>The local path of the CSV file where the budgets were written.</returns>
-	public string WriteBudgetsToCsvFile(List<GetBudgetsToExportInfo> budgets, string filePath)
-	{
-		// Write text to the file
-		using (var writer = new StreamWriter(filePath))
-		using (var csv = new CsvWriter(writer, CultureInfo.InvariantCulture))
-		{
-			csv.WriteHeader<GetBudgetsToExportInfo>();
-			csv.NextRecord();
-			foreach (var budget in budgets)
-			{
-				csv.WriteRecord(budget);
-				csv.NextRecord();
-			}
-		}
-
-		return filePath;
-	}
-
-	/// <summary>
-	/// Uploads a CSV file containing a list of budgets to Azure Blob Storage.
-	/// </summary>
-	/// <param name="budgetInfos">A list of budgets to be written to the CSV file and uploaded.</param>
-	/// <param name="containerClient">Client for interacting with a specific blob container in Azure Blob Storage.</param>
-	/// <returns>The absolute URI of the uploaded blob in Azure Blob Storage.</returns>
-	public async Task<string> UploadToBlobStorage(List<GetBudgetsToExportInfo> budgetInfos, BlobContainerClient containerClient)
-	{
-		string localFilePath = this.GenerateLocalCsvFilePath();
-		string filePath = this.WriteBudgetsToCsvFile(budgetInfos, localFilePath);
-
-		BlobClient blobClient = containerClient.GetBlobClient(Path.GetFileName(filePath));
-
-		await blobClient.UploadAsync(filePath, true);
-		File.Delete(filePath);
-
-		return blobClient.Uri.AbsoluteUri;
-	}
-
-	/// <summary>
-	/// Downloads a CSV file containing a list of budgets from Azure Blob Storage and imports the budgets into the application.
-	/// </summary>
-	/// <param name="filename">The name of the blob to be downloaded from Azure Blob Storage.</param>
-	/// <param name="containerClient">Client for interacting with a specific blob container in Azure Blob Storage.</param>
-	/// <param name="csvConfig">Configuration for reading the CSV file.</param>
-	/// <returns>A task that represents the asynchronous operation.</returns>
-	public async Task ImportBudgetsFromBlobStorage(string filename, BlobContainerClient containerClient, CsvConfiguration csvConfig)
-	{
-		BlobClient blobClient = containerClient.GetBlobClient(Path.GetFileName(filename));
-		BlobDownloadInfo download = await blobClient.DownloadAsync();
-
-		using (var reader = new StreamReader(download.Content))
-		{
-			using (var csv = new CsvReader(reader, csvConfig))
-			{
-				csv.Read();
-				var budgetsToImport = csv.GetRecords<GetBudgetsToExportInfo>();
-
-				foreach (var budget in budgetsToImport)
-				{
-					var userId = this.contextAccessor.GetUserId()!.Value;
-					decimal limit = decimal.Parse(budget.Value, CultureInfo.InvariantCulture);
-					var money = new Money(limit, (Currency)Enum.Parse(typeof(Currency), budget.Currency));
-					var period = new Period(DateTime.Parse(budget.StartDate), DateTime.Parse(budget.EndDate));
-					string description = string.IsNullOrEmpty(budget.Description) ? string.Empty : budget.Description;
-
-					var newbudget = new CreateBudget(Guid.NewGuid(), budget.Name, userId, money, period, description, budget.IconName);
-
-					await this.commandBus.Send(newbudget);
-				}
-			}
-		}
-	}
-
-	/// <summary>
-	/// Reads budgets from a CSV file, validates them, and returns a list of valid budgets.
-	/// Any errors encountered during validation are added to the provided errors list.
-	/// </summary>
-	/// <param name="file">The CSV file containing the budgets to be read and validated.</param>
-	/// <param name="csvConfig">Configuration for reading the CSV file.</param>
-	/// <param name="errors">A list to which any validation errors will be added.</param>
-	/// <returns>A list of valid budgets read from the CSV file.</returns>
-	public List<GetBudgetsToExportInfo> ReadAndValidateBudgets(IFormFile file, CsvConfiguration csvConfig, List<string> errors)
-	{
-		var budgetInfos = new List<GetBudgetsToExportInfo>();
-		using var stream = file.OpenReadStream();
-		using (var csv = new CsvReader(new StreamReader(stream), csvConfig))
-		{
-			csv.Read();
-			var budgets = csv.GetRecords<GetBudgetsToExportInfo>().ToList();
-			int rowNumber = 0;
-
-			foreach (var budget in budgets)
-			{
-				var results = this.ValidateBudget(budget);
-				rowNumber++;
-
-				if (results.Any())
-				{
-					foreach (string result in results)
-					{
-						errors.Add($"row: {rowNumber}| error: {result}");
-					}
-
-					continue;
-				}
-
-				var updateBudget = this.CreateBudgetInfoAsync(budget);
-				budgetInfos.Add(updateBudget!);
-			}
-		}
-
-		return budgetInfos;
-	}
-
-	/// <summary>
-	/// Creates a new budget based on the provided budget information.
-	/// If a budget with the same name already exists in the database, a random number is appended to the name.
-	/// </summary>
-	/// <param name="budget">The budget information used to create the new budget.</param>
-	/// <returns>Creates a new budget.</returns>
-	public GetBudgetsToExportInfo? CreateBudgetInfoAsync(GetBudgetsToExportInfo budget)
-	{
-		bool isExistingBudget = this.budgetDbContext.Budget.Any(b => b.Name.Equals(budget.Name));
-		string budgetName = isExistingBudget ? budget.Name + new Random().Next(100000, 900001) : budget.Name;
-
-		var budgetInfo = new GetBudgetsToExportInfo
-		{
-			Name = budgetName,
-			IconName = budget.IconName,
-			Description = budget.Description,
-			Currency = budget.Currency,
-			Value = budget.Value,
-			StartDate = budget.StartDate,
-			EndDate = budget.EndDate,
-		};
-
-		return budgetInfo;
-	}
-}
